Module structopt_toml::serde

Re-export of serde

Modules

de

Generic data structure deserialization framework.

ser

Generic data structure serialization framework.

Macros

forward_to_deserialize_any

Helper macro when implementing the Deserializer part of a new data format for Serde.

serde_if_integer128

Conditional compilation depending on whether Serde is built with support for 128-bit integers.

Traits

Deserialize

A data structure that can be deserialized from any data format supported by Serde.

Deserializer

A data format that can deserialize any data structure supported by Serde.

Serialize

A data structure that can be serialized into any data format supported by Serde.

Serializer

A data format that can serialize any data structure supported by Serde.

Derive Macros

Deserialize

Serialize